refactor: Separate multiple internal classes from single file
- Address multiple todos that mostly don't affect the API too much
